Split up tests for s3 resources plugin and upload
